+++ src/tables/mod.rs
@@ -23,5 +23,6 @@ pub mod maxp;
 pub mod name;
 pub mod os2;
 pub mod post;
+pub mod stat;
 pub mod vhea;
 pub mod vmtx;